# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  model.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  model.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@anixops.com
    name: API Support
    url: https://github.com/AnixOps/anix-control/issues
  description: |-
    AnixOps Control 统一控制面 API 文档
    支持用户管理、节点管理、订阅系统、支付网关、流量转发等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/AnixOps/anix-control
  title: AnixOps Control API
  version: 3.0.0-alpha.1
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: 用户登录获取 JWT Token
      parameters:
      - description: 登录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 认证失败
          schema:
            additionalProperties: true
            type: object
      summary: 用户登录
      tags:
      - 认证
...
  /register:
    post:
      consumes:
      - application/json
      description: 创建新用户账户
      parameters:
      - description: 注册请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      summary: 用户注册
      tags:
      - 认证
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...